% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{create_task}
\alias{create_task}
\title{library(caret)
library(mlr3pipelines)
library(mlr3)
library(mlr3tuning)
library(paradox)}
\usage{
create_task(abt, target = "Class", test_oot = FALSE,
  label_cols = c("cpr_nr", "se_nr", "rating"))
}
\arguments{
\item{abt}{\code{data.frame}}

\item{target}{\code{character} name of target variable.}
}
\value{
\code{task}.
}
\description{
# load data.
data(GermanCredit)
abt <- GermanCredit
Create Task for Compliance Training
}
