% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{create_tuner}
\alias{create_tuner}
\title{Create Tuner for Model Training}
\usage{
create_tuner(GL, validation_scheme = mlr_resamplings$get("cv"),
  measure = mlr_measures$mget("classif.acc"))
}
\arguments{
\item{validation_scheme}{}
}
\value{
tuner.
}
\description{
Create Tuner for Model Training
}
