% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_model}
\alias{train_model}
\title{Title}
\usage{
train_model(abt, col_target = "Class", col_labels = c("cpr_nr",
  "se_nr", "rating"), test_oot = FALSE,
  validation_scheme = mlr_resamplings$get("cv"),
  measure = mlr_measures$mget("classif.acc"),
  model = "complianceBorger")
}
\description{
Title
}
